refactor(asserts): step 0: scaffold tests/internal/asserts and inline single-callers
Add tests/internal/asserts package skeleton with a package doc.go. Remove the
unused AssertClusterRollingRestart helper and relocate six helpers with
a single external caller from tests/e2e/asserts_test.go to the test
files that use them: OnlineResizePVC/OfflineResizePVC to
storage_expansion_test.go, DeleteTableUsingPgBouncerService to
pgbouncer_test.go, AssertWebhookEnabled to webhook_test.go,
AssertClusterAsyncReplica to backup_restore_minio_test.go, and
AssertBackupConditionTimestampChangedInClusterStatus to
tablespaces_test.go.

No behaviour change; this is a preparation step for moving the asserts
into importable per-topic packages under tests/internal/asserts/.
