+++ tests/e2e/asserts_test.go
@@ -58,7 +58,6 @@ import (
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/minio"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/nodes"
 	objectsutils "github.com/cloudnative-pg/cloudnative-pg/tests/utils/objects"
-	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/operator"
 	podutils "github.com/cloudnative-pg/cloudnative-pg/tests/utils/pods"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/postgres"
 	"github.com/cloudnative-pg/cloudnative-pg/tests/utils/proxy"
@@ -414,24 +413,6 @@ func AssertClusterDefault(
 		validationWarn, validationErr := validator.ValidateCreate(env.Ctx, cluster)
 		Expect(validationWarn).To(BeEmpty())
 		Expect(validationErr).ToNot(HaveOccurred())
-	})
-}
-
-func AssertWebhookEnabled(env *environment.TestingEnvironment, mutating, validating string) {
-	By("re-setting namespace selector for all admission controllers", func() {
-		// Setting the namespace selector in MutatingWebhook and ValidatingWebhook
-		// to nil will go back to the default behaviour
-		mWhc, position, err := operator.GetMutatingWebhookByName(env.Ctx, env.Client, mutating)
-		Expect(err).ToNot(HaveOccurred())
-		mWhc.Webhooks[position].NamespaceSelector = nil
-		err = operator.UpdateMutatingWebhookConf(env.Ctx, env.Interface, mWhc)
-		Expect(err).ToNot(HaveOccurred())
-
-		vWhc, position, err := operator.GetValidatingWebhookByName(env.Ctx, env.Client, validating)
-		Expect(err).ToNot(HaveOccurred())
-		vWhc.Webhooks[position].NamespaceSelector = nil
-		err = operator.UpdateValidatingWebhookConf(env.Ctx, env.Interface, vWhc)
-		Expect(err).ToNot(HaveOccurred())
 	})
 }
 
@@ -1629,94 +1610,6 @@ func AssertSSLVerifyFullDBConnectionFromAppPod(namespace string, clusterName str
 	})
 }
 
-func AssertClusterAsyncReplica(namespace, sourceClusterFile, restoreClusterFile, tableName string) {
-	By("Async Replication into external cluster", func() {
-		restoredClusterName, err := yaml.GetResourceNameFromYAML(env.Scheme, restoreClusterFile)
-		Expect(err).ToNot(HaveOccurred())
-		// Add additional data to the source cluster
-		sourceClusterName, err := yaml.GetResourceNameFromYAML(env.Scheme, sourceClusterFile)
-		Expect(err).ToNot(HaveOccurred())
-		CreateResourceFromFile(namespace, restoreClusterFile)
-		// We give more time than the usual 600s, since the recovery is slower
-		AssertClusterIsReady(namespace, restoredClusterName, testTimeouts[timeouts.ClusterIsReadySlow], env)
-
-		// Test data should be present on restored primary
-		restoredPrimary, err := clusterutils.GetPrimary(env.Ctx, env.Client, namespace, restoredClusterName)
-		Expect(err).ToNot(HaveOccurred())
-
-		// We need the credentials from the source cluster because the replica cluster
-		// doesn't create the credentials on its own namespace
-		appUser, appUserPass, err := secrets.GetCredentials(
-			env.Ctx,
-			env.Client,
-			sourceClusterName,
-			namespace,
-			apiv1.ApplicationUserSecretSuffix,
-		)
-		Expect(err).ToNot(HaveOccurred())
-
-		forwardRestored, connRestored, err := postgres.ForwardPSQLConnectionWithCreds(
-			env.Ctx,
-			env.Client,
-			env.Interface,
-			env.RestClientConfig,
-			namespace,
-			restoredClusterName,
-			postgres.AppDBName,
-			appUser,
-			appUserPass,
-		)
-		defer func() {
-			_ = connRestored.Close()
-			forwardRestored.Close()
-		}()
-		Expect(err).ToNot(HaveOccurred())
-
-		row := connRestored.QueryRow(fmt.Sprintf("SELECT count(*) FROM %s", tableName))
-		var countString string
-		err = row.Scan(&countString)
-		Expect(err).ToNot(HaveOccurred())
-		Expect(countString).To(BeEquivalentTo("2"))
-
-		forwardSource, connSource, err := postgres.ForwardPSQLConnection(
-			env.Ctx,
-			env.Client,
-			env.Interface,
-			env.RestClientConfig,
-			namespace,
-			sourceClusterName,
-			postgres.AppDBName,
-			apiv1.ApplicationUserSecretSuffix,
-		)
-		defer func() {
-			_ = connSource.Close()
-			forwardSource.Close()
-		}()
-		Expect(err).ToNot(HaveOccurred())
-
-		// Insert new data in the source cluster
-		insertRecordIntoTable(tableName, 3, connSource)
-		AssertArchiveWalOnMinio(namespace, sourceClusterName, sourceClusterName)
-		tableLocator := TableLocator{
-			Namespace:    namespace,
-			ClusterName:  sourceClusterName,
-			DatabaseName: postgres.AppDBName,
-			TableName:    tableName,
-		}
-		AssertDataExpectedCount(env, tableLocator, 3)
-
-		cluster, err := clusterutils.Get(env.Ctx, env.Client, namespace, restoredClusterName)
-		Expect(err).ToNot(HaveOccurred())
-		expectedReplicas := cluster.Spec.Instances - 1
-		// Cascading replicas should be attached to primary replica
-		connectedReplicas, err := postgres.CountReplicas(
-			env.Ctx, env.Client, env.Interface, env.RestClientConfig,
-			restoredPrimary, RetryTimeout,
-		)
-		Expect(connectedReplicas, err).To(BeEquivalentTo(expectedReplicas))
-	})
-}
-
 func AssertClusterRestoreWithApplicationDB(namespace, restoreClusterFile, tableName string) {
 	restoredClusterName, err := yaml.GetResourceNameFromYAML(env.Scheme, restoreClusterFile)
 	Expect(err).ToNot(HaveOccurred())
@@ -2370,202 +2263,6 @@ func assertPGBouncerHasServiceNameInsideHostParameter(namespace, serviceName str
 	}
 }
 
-// OnlineResizePVC is for verifying if storage can be automatically expanded, or not
-func OnlineResizePVC(namespace, clusterName string) {
-	walStorageEnabled, err := storage.IsWalStorageEnabled(
-		env.Ctx, env.Client,
-		namespace, clusterName,
-	)
-	Expect(err).ToNot(HaveOccurred())
-
-	pvc := &corev1.PersistentVolumeClaimList{}
-	By("verify PVC before expansion", func() {
-		// Verifying the first stage of deployment to compare it further with expanded value
-		err := env.Client.List(env.Ctx, pvc, ctrlclient.InNamespace(namespace))
-		Expect(err).ToNot(HaveOccurred())
-		// Iterating through PVC list to assure its default size
-		for _, pvClaim := range pvc.Items {
-			Expect(pvClaim.Status.Capacity.Storage().String()).To(BeEquivalentTo("1Gi"))
-		}
-	})
-	By("expanding Cluster storage", func() {
-		// Patching cluster to expand storage size from 1Gi to 2Gi
-		storageType := []string{"storage"}
-		if walStorageEnabled {
-			storageType = append(storageType, "walStorage")
-		}
-		for _, s := range storageType {
-			cmd := fmt.Sprintf(
-				"kubectl patch cluster %v -n %v -p '{\"spec\":{\"%v\":{\"size\":\"2Gi\"}}}' --type=merge",
-				clusterName,
-				namespace,
-				s)
-			Eventually(func() error {
-				_, _, err := run.Unchecked(cmd)
-				return err
-			}, 60, 5).Should(Succeed())
-		}
-	})
-	By("verifying Cluster storage is expanded", func() {
-		// Gathering and verifying the new size of PVC after update on cluster
-		expectedCount := 3
-		if walStorageEnabled {
-			expectedCount = 6
-		}
-		Eventually(func(g Gomega) {
-			// Variable counter to store the updated total of expanded PVCs. It should be equal to three
-			updateCount := 0
-			// Gathering PVC list
-			err := env.Client.List(env.Ctx, pvc, ctrlclient.InNamespace(namespace))
-			g.Expect(err).ToNot(HaveOccurred())
-			// Iterating through PVC list to compare with expanded size
-			for _, pvClaim := range pvc.Items {
-				// Size comparison
-				if pvClaim.Status.Capacity.Storage().String() == "2Gi" {
-					updateCount++
-				}
-			}
-			g.Expect(updateCount).To(BeEquivalentTo(expectedCount))
-		}, 300).Should(Succeed())
-	})
-}
-
-func OfflineResizePVC(namespace, clusterName string, timeout int) {
-	walStorageEnabled, err := storage.IsWalStorageEnabled(
-		env.Ctx, env.Client,
-		namespace, clusterName,
-	)
-	Expect(err).ToNot(HaveOccurred())
-
-	By("verify PVC size before expansion", func() {
-		// Gathering PVC list for future use of comparison and deletion after storage expansion
-		pvc := &corev1.PersistentVolumeClaimList{}
-		err := env.Client.List(env.Ctx, pvc, ctrlclient.InNamespace(namespace))
-		Expect(err).ToNot(HaveOccurred())
-		// Iterating through PVC list to verify the default size for future comparison
-		for _, pvClaim := range pvc.Items {
-			Expect(pvClaim.Status.Capacity.Storage().String()).To(BeEquivalentTo("1Gi"))
-		}
-	})
-	By("expanding Cluster storage", func() {
-		// Expanding cluster storage
-		storageType := []string{"storage"}
-		if walStorageEnabled {
-			storageType = append(storageType, "walStorage")
-		}
-		for _, s := range storageType {
-			cmd := fmt.Sprintf(
-				"kubectl patch cluster %v -n %v -p '{\"spec\":{\"%v\":{\"size\":\"2Gi\"}}}' --type=merge",
-				clusterName,
-				namespace,
-				s)
-			Eventually(func() error {
-				_, _, err := run.Unchecked(cmd)
-				return err
-			}, 60, 5).Should(Succeed())
-		}
-	})
-	By("deleting Pod and PVCs, first replicas then the primary", func() {
-		// Gathering cluster primary
-		currentPrimary, err := clusterutils.GetPrimary(env.Ctx, env.Client, namespace, clusterName)
-		Expect(err).ToNot(HaveOccurred())
-		currentPrimaryWalStorageName := currentPrimary.Name + "-wal"
-		quickDelete := &ctrlclient.DeleteOptions{
-			GracePeriodSeconds: &quickDeletionPeriod,
-		}
-
-		podList, err := clusterutils.ListPods(env.Ctx, env.Client, namespace, clusterName)
-		Expect(err).ToNot(HaveOccurred())
-		Expect(len(podList.Items), err).To(BeEquivalentTo(3))
-
-		// Iterating through PVC list for deleting pod and PVC for storage expansion
-		for _, p := range podList.Items {
-			// Comparing cluster pods to not be primary to ensure cluster is healthy.
-			// Primary will be eventually deleted
-			if !specs.IsPodPrimary(p) {
-				// Deleting PVC
-				_, _, err = run.Run(
-					"kubectl delete pvc " + p.Name + " -n " + namespace + " --wait=false")
-				Expect(err).ToNot(HaveOccurred())
-				// Deleting WalStorage PVC if needed
-				if walStorageEnabled {
-					_, _, err = run.Run(
-						"kubectl delete pvc " + p.Name + "-wal" + " -n " + namespace + " --wait=false")
-					Expect(err).ToNot(HaveOccurred())
-				}
-				// Deleting standby and replica pods
-				err = podutils.Delete(env.Ctx, env.Client, namespace, p.Name, quickDelete)
-				Expect(err).ToNot(HaveOccurred())
-			}
-		}
-		AssertClusterIsReady(namespace, clusterName, timeout, env)
-
-		// Deleting primary pvc
-		_, _, err = run.Run(
-			"kubectl delete pvc " + currentPrimary.Name + " -n " + namespace + " --wait=false")
-		Expect(err).ToNot(HaveOccurred())
-		// Deleting Primary WalStorage PVC if needed
-		if walStorageEnabled {
-			_, _, err = run.Run(
-				"kubectl delete pvc " + currentPrimaryWalStorageName + " -n " + namespace + " --wait=false")
-			Expect(err).ToNot(HaveOccurred())
-		}
-		// Deleting primary pod
-		err = podutils.Delete(env.Ctx, env.Client, namespace, currentPrimary.Name, quickDelete)
-		Expect(err).ToNot(HaveOccurred())
-	})
-
-	AssertClusterIsReady(namespace, clusterName, timeout, env)
-	By("verifying Cluster storage is expanded", func() {
-		// Gathering PVC list for comparison
-		pvcList, err := storage.GetPVCList(env.Ctx, env.Client, namespace)
-		Expect(err).ToNot(HaveOccurred())
-		// Gathering PVC size and comparing with expanded value
-		expectedCount := 3
-		if walStorageEnabled {
-			expectedCount = 6
-		}
-		Eventually(func() int {
-			// Bool value to ensure every pod in cluster expanded, will be eventually compared as true
-			count := 0
-			// Iterating through PVC list for comparison
-			for _, pvClaim := range pvcList.Items {
-				// Comparing to expanded value.
-				// Once all pods will be expanded, count will be equal to three
-				if pvClaim.Status.Capacity.Storage().String() == "2Gi" {
-					count++
-				}
-			}
-			return count
-		}, 30).Should(BeEquivalentTo(expectedCount))
-	})
-}
-
-func DeleteTableUsingPgBouncerService(
-	namespace,
-	clusterName,
-	poolerYamlFilePath string,
-	env *environment.TestingEnvironment,
-	pod *corev1.Pod,
-) {
-	poolerService, err := yaml.GetResourceNameFromYAML(env.Scheme, poolerYamlFilePath)
-	Expect(err).ToNot(HaveOccurred())
-
-	appUser, generatedAppUserPassword, err := secrets.GetCredentials(
-		env.Ctx, env.Client,
-		clusterName, namespace, apiv1.ApplicationUserSecretSuffix,
-	)
-	Expect(err).ToNot(HaveOccurred())
-	AssertConnection(namespace, poolerService, postgres.AppDBName, appUser, generatedAppUserPassword, env)
-
-	connectionTimeout := time.Second * 10
-	dsn := services.CreateDSN(poolerService, appUser, postgres.AppDBName, generatedAppUserPassword,
-		services.Require, 5432)
-	_, _, err = env.EventuallyExecCommand(env.Ctx, *pod, specs.PostgresContainerName, &connectionTimeout,
-		"psql", dsn, "-tAc", "DROP TABLE table1")
-	Expect(err).ToNot(HaveOccurred())
-}
-
 func collectAndAssertDefaultMetricsPresentOnEachPod(
 	namespace, clusterName string,
 	tlsEnabled bool,
@@ -2782,25 +2479,6 @@ func DeleteResourcesFromFile(namespace, sampleFilePath string) error {
 	return nil
 }
 
-func AssertBackupConditionTimestampChangedInClusterStatus(
-	namespace,
-	clusterName string,
-	clusterConditionType apiv1.ClusterConditionType,
-	lastTransactionTimeStamp *metav1.Time,
-) {
-	By(fmt.Sprintf("waiting for backup condition status in cluster '%v'", clusterName), func() {
-		Eventually(func() (bool, error) {
-			getBackupCondition, err := backups.GetConditionsInClusterStatus(
-				env.Ctx, env.Client,
-				namespace, clusterName, clusterConditionType)
-			if err != nil {
-				return false, err
-			}
-			return getBackupCondition.LastTransitionTime.After(lastTransactionTimeStamp.Time), nil
-		}, 300, 5).Should(BeTrue())
-	})
-}
-
 func AssertClusterReadinessStatusIsReached(
 	namespace,
 	clusterName string,
@@ -2967,25 +2645,6 @@ func AssertClusterHAReplicationSlots(namespace, clusterName string) {
 		}
 		AssertClusterReplicationSlotsAligned(namespace, clusterName)
 	})
-}
-
-// AssertClusterRollingRestart restarts a given cluster
-func AssertClusterRollingRestart(namespace, clusterName string) {
-	By(fmt.Sprintf("restarting cluster %v", clusterName), func() {
-		cluster, err := clusterutils.Get(env.Ctx, env.Client, namespace, clusterName)
-		Expect(err).ToNot(HaveOccurred())
-		clusterRestarted := cluster.DeepCopy()
-		if clusterRestarted.Annotations == nil {
-			clusterRestarted.Annotations = make(map[string]string)
-		}
-		clusterRestarted.Annotations[utils.ClusterRestartAnnotationName] = time.Now().Format(time.RFC3339)
-		clusterRestarted.ManagedFields = nil
-		err = env.Client.Patch(env.Ctx, clusterRestarted, ctrlclient.MergeFrom(cluster))
-		Expect(err).ToNot(HaveOccurred())
-	})
-	AssertClusterEventuallyReachesPhase(namespace, clusterName,
-		[]string{apiv1.PhaseUpgrade, apiv1.PhaseWaitingForInstancesToBeActive}, 120)
-	AssertClusterIsReady(namespace, clusterName, testTimeouts[timeouts.ClusterIsReadyQuick], env)
 }
 
 // AssertPVCCount matches count and pvc List.
